requirements.txt
Flask==3.0.0
tensorflow==2.15.0
tensorflow-datasets==4.9.5
Pillow==9.3.0
